requirements.txt
djangorestframework
django
pyreportjasper
psycopg2
